Show orbital regimes on Sol fact sheet
